refactor: codebase health round 6 - split large Python files
Extracted 4 new focused modules from 2 large files:

- ability_handlers.py (822→90 LOC): extracted ability_handlers_summon.py (262),
  ability_handlers_elemental.py (252), ability_handlers_physical.py (236)
- battle_manager.py (776→491 LOC): extracted battle_player_actions.py (359)

Engine.py skipped as it was already refactored in Round 4.

Total reduction: 1598 → 581 LOC in target files (64% reduction)
