+++ docs/refactoring/round6-baseline.md
@@ -1,0 +1,34 @@
+# Refactoring Round 6 - Baseline
+
+Generated: 2026-01-13
+Starting from Round 5 completion
+Target: Split remaining Python files over 700 LOC
+
+## Round 6 Targets
+
+### Python Core (non-data files over 700 LOC)
+| File | LOC | Technique |
+|------|-----|-----------|
+| ability_handlers.py | 822 | Extract by ability category |
+| engine.py | 797 | Already refactored in Round 4 - skip |
+| battle_manager.py | 776 | Extract phase handlers, turn logic |
+
+### Deferred (data-heavy or lower priority)
+| File | LOC | Notes |
+|------|-----|-------|
+| ghosts/manager.py | 710 | Ghost system - complex state |
+| ai_scoring.py | 700 | AI logic - tightly coupled |
+| artifacts.py | 682 | Data file - DB migration candidate |
+| story_data.py | 638 | Data file - DB migration candidate |
+
+## Refactoring Log
+
+| Date | File | Before LOC | After LOC | Changes Made |
+|------|------|------------|-----------|--------------|
+| 2026-01-13 | ability_handlers.py | 822 | 90 | Extracted ability_handlers_summon.py (262), ability_handlers_elemental.py (252), ability_handlers_physical.py (236) |
+| 2026-01-13 | battle_manager.py | 776 | 491 | Extracted battle_player_actions.py (359) |
+
+## Summary
+
+Round 6 reduced 2 files from 1598 total LOC to 581 LOC (64% reduction in target files).
+Created 4 new focused modules. Engine.py skipped (already refactored in Round 4).
