+++ src/entities/ability_handlers.py
@@ -1,790 +1,58 @@
 """Boss ability execution handlers.
 
-Contains all _execute_* functions for individual abilities.
+Contains the dispatch table mapping ability names to handler functions.
+Individual handlers are organized by category in separate modules.
 """
-from typing import TYPE_CHECKING, Tuple
-import random
-
-from .ability_definitions import BossAbility
-
-if TYPE_CHECKING:
-    from .entities import Enemy, Player
-    from ..world.dungeon import Dungeon
-    from ..managers.entity_manager import EntityManager
-
-
-def calculate_elemental_damage(
-    base_damage: int,
-    element_type: str,
-    target,
-    attacker=None,
-) -> Tuple[int, str]:
-    """
-    Calculate damage with elemental weaknesses and resistances.
-
-    Args:
-        base_damage: Base damage before modifiers
-        element_type: Element of the attack ('fire', 'ice', 'lightning', 'dark')
-        target: Entity being attacked (has resistances dict and current_element)
-        attacker: Entity attacking (for weakness checking)
-
-    Returns:
-        Tuple of (final_damage, modifier_message)
-    """
-    from ..core.constants import (
-        ElementType, ELEMENT_WEAKNESSES, WEAKNESS_DAMAGE_MULTIPLIER
-    )
-
-    damage = base_damage
-    message = ""
-
-    # Get target's resistances
-    resistances = getattr(target, 'resistances', {})
-    resistance = resistances.get(element_type, 0.0)
-
-    # Apply resistance reduction
-    if resistance >= 1.0:
-        # Immune - no damage
-        return 0, " (Immune!)"
-    elif resistance > 0:
-        # Partial resistance
-        damage = int(damage * (1.0 - resistance))
-        message = f" (Resisted {int(resistance * 100)}%)"
-
-    # Check for weakness (target's current element is weak to attack element)
-    target_element = getattr(target, 'current_element', None)
-    if target_element:
-        # Map attack element string to ElementType
-        attack_element_map = {
-            'fire': ElementType.FIRE,
-            'ice': ElementType.ICE,
-            'lightning': ElementType.LIGHTNING,
-            'dark': ElementType.DARK,
-        }
-        attack_element = attack_element_map.get(element_type)
-
-        # Check if attack element is what target is weak to
-        if attack_element:
-            weakness = ELEMENT_WEAKNESSES.get(target_element)
-            if weakness == attack_element:
-                # Super effective!
-                damage = int(damage * WEAKNESS_DAMAGE_MULTIPLIER)
-                message = " (Super effective!)"
-
-    return max(1, damage), message
-
-
-# =============================================================================
-# Original Boss Ability Handlers (v1.0-v3.0)
-# =============================================================================
-
-def execute_summon_goblins(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summon 2-3 goblin minions near the boss."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    num_goblins = random.randint(2, 3)
-    spawned = 0
-
-    # Find valid spawn positions adjacent to boss
-    for _ in range(num_goblins):
-        for dx in range(-2, 3):
-            for dy in range(-2, 3):
-                if dx == 0 and dy == 0:
-                    continue
-                nx, ny = boss.x + dx, boss.y + dy
-                if (dungeon.is_walkable(nx, ny) and
-                    not entity_manager.get_enemy_at(nx, ny) and
-                    (nx != player.x or ny != player.y)):
-                    goblin = Enemy(nx, ny, enemy_type=EnemyType.GOBLIN, is_elite=False)
-                    entity_manager.enemies.append(goblin)
-                    spawned += 1
-                    break
-            else:
-                continue
-            break
-
-    if spawned > 0:
-        return True, f"The {boss.name} summons {spawned} goblins!", 0
-    return False, "", 0
-
-
-def execute_war_cry(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Buff boss damage for 3 turns."""
-    boss.buff_turns['war_cry'] = 3
-    boss.buffed_damage = int(boss.base_attack_damage * 1.5)
-    return True, f"The {boss.name} lets out a terrifying war cry! (+50% damage)", 0
-
-
-def execute_ground_slam(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """AOE attack hitting all within range."""
-    distance = abs(player.x - boss.x) + abs(player.y - boss.y)
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        return True, f"The {boss.name} slams the ground! You take {damage} damage!", damage
-    return True, f"The {boss.name} slams the ground, but you're out of range!", 0
-
-
-def execute_regenerate(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Heal when below 50% HP."""
-    if boss.health < boss.max_health // 2:
-        heal_amount = min(-ability.damage, boss.max_health - boss.health)
-        boss.health += heal_amount
-        return True, f"The {boss.name} regenerates {heal_amount} health!", 0
-    return False, "", 0
-
-
-def execute_raise_dead(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summon 2 skeleton minions."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    spawned = 0
-    for _ in range(2):
-        for dx in range(-2, 3):
-            for dy in range(-2, 3):
-                if dx == 0 and dy == 0:
-                    continue
-                nx, ny = boss.x + dx, boss.y + dy
-                if (dungeon.is_walkable(nx, ny) and
-                    not entity_manager.get_enemy_at(nx, ny) and
-                    (nx != player.x or ny != player.y)):
-                    skeleton = Enemy(nx, ny, enemy_type=EnemyType.SKELETON, is_elite=False)
-                    entity_manager.enemies.append(skeleton)
-                    spawned += 1
-                    break
-            else:
-                continue
-            break
-
-    if spawned > 0:
-        return True, f"The {boss.name} raises {spawned} skeletons from the dead!", 0
-    return False, "", 0
-
-
-def execute_life_drain(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Drain life from player, healing self."""
-    distance = abs(player.x - boss.x) + abs(player.y - boss.y)
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        heal = min(damage, boss.max_health - boss.health)
-        boss.health += heal
-        return True, f"The {boss.name} drains your life for {damage} damage, healing {heal}!", damage
-    return False, "", 0
-
-
-def execute_arcane_bolt(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Ranged attack."""
-    distance = abs(player.x - boss.x) + abs(player.y - boss.y)
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        return True, f"The {boss.name} fires an arcane bolt for {damage} damage!", damage
-    return False, "", 0
-
-
-def execute_teleport(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Teleport to a random location."""
-    # Try to find a valid teleport location
-    for _ in range(50):
-        x, y = dungeon.get_random_floor_position()
-        if (dungeon.is_walkable(x, y) and
-            not entity_manager.get_enemy_at(x, y) and
-            (x != player.x or y != player.y)):
-            boss.x = x
-            boss.y = y
-            return True, f"The {boss.name} vanishes and reappears elsewhere!", 0
-    return False, "", 0
-
-
-def execute_fire_breath(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Cone AOE attack toward player."""
-    distance = abs(player.x - boss.x) + abs(player.y - boss.y)
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        return True, f"The {boss.name} breathes fire for {damage} damage!", damage
-    return True, f"The {boss.name} breathes fire, but you dodge!", 0
-
-
-def execute_tail_sweep(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Melee AOE hitting all adjacent."""
-    distance = max(abs(player.x - boss.x), abs(player.y - boss.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        return True, f"The {boss.name} sweeps its tail for {damage} damage!", damage
-    return False, "", 0
-
-
-# =============================================================================
-# v4.0 New Ability Handlers
-# =============================================================================
-
-def execute_raise_skeleton(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summon a single skeleton minion."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    # Find a valid spawn position near the caster
-    for dx in range(-2, 3):
-        for dy in range(-2, 3):
-            if dx == 0 and dy == 0:
-                continue
-            nx, ny = caster.x + dx, caster.y + dy
-            if (dungeon.is_walkable(nx, ny) and
-                not entity_manager.get_enemy_at(nx, ny) and
-                (nx != player.x or ny != player.y)):
-                skeleton = Enemy(nx, ny, enemy_type=EnemyType.SKELETON, is_elite=False)
-                entity_manager.enemies.append(skeleton)
-                return True, f"The {caster.name} raises a skeleton from the shadows!", 0
-
-    return False, "", 0
-
-
-def execute_dark_bolt(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Ranged dark magic attack."""
-    distance = abs(player.x - caster.x) + abs(player.y - caster.y)
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        return True, f"The {caster.name} fires a dark bolt for {damage} damage!", damage
-    return False, "", 0
-
-
-def execute_fire_strike(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Melee attack that applies burn status."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        burn_msg = player.apply_status_effect(StatusEffectType.BURN, caster.name)
-        return True, f"The {caster.name} strikes with fire for {damage} damage! {burn_msg}", damage
-    return False, "", 0
-
-
-def execute_backstab(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """3x damage attack from stealth."""
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        # Check if caster is invisible (stealthed)
-        is_stealthed = getattr(caster, 'is_invisible', False)
-        multiplier = 3 if is_stealthed else 1
-        base_damage = caster.attack_damage * multiplier
-        damage = player.take_damage(base_damage)
-
-        # Break stealth after attacking
-        if is_stealthed:
-            caster.is_invisible = False
-            return True, f"The {caster.name} backstabs you for {damage} critical damage!", damage
-
-        return True, f"The {caster.name} strikes for {damage} damage!", damage
-    return False, "", 0
-
-
-def execute_vanish(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Enter invisibility."""
-    caster.is_invisible = True
-    return True, f"The {caster.name} vanishes into the shadows!", 0
-
-
-def execute_fire_bolt(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Ranged fire attack that applies burn."""
-    from ..core.constants import StatusEffectType
-
-    distance = abs(player.x - caster.x) + abs(player.y - caster.y)
-    if distance <= ability.range:
-        # Apply elemental damage calculation
-        final_damage, element_msg = calculate_elemental_damage(ability.damage, 'fire', player, caster)
-        damage = player.take_damage(final_damage)
-        burn_msg = player.apply_status_effect(StatusEffectType.BURN, caster.name)
-        return True, f"The {caster.name} hurls a fire bolt for {damage} damage{element_msg}! {burn_msg}", damage
-    return False, "", 0
-
-
-def execute_ice_shard(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Ranged ice attack that applies freeze."""
-    from ..core.constants import StatusEffectType
-
-    distance = abs(player.x - caster.x) + abs(player.y - caster.y)
-    if distance <= ability.range:
-        # Apply elemental damage calculation
-        final_damage, element_msg = calculate_elemental_damage(ability.damage, 'ice', player, caster)
-        damage = player.take_damage(final_damage)
-        freeze_msg = player.apply_status_effect(StatusEffectType.FREEZE, caster.name)
-        return True, f"The {caster.name} launches an ice shard for {damage} damage{element_msg}! {freeze_msg}", damage
-    return False, "", 0
-
-
-def execute_chain_lightning(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Lightning that can jump to nearby enemies (damages player primarily)."""
-    distance = abs(player.x - caster.x) + abs(player.y - caster.y)
-    if distance <= ability.range:
-        # Apply elemental damage calculation
-        final_damage, element_msg = calculate_elemental_damage(ability.damage, 'lightning', player, caster)
-        damage = player.take_damage(final_damage)
-        msg = f"The {caster.name} unleashes chain lightning for {damage} damage{element_msg}!"
-
-        # Lightning can stun
-        from ..core.constants import StatusEffectType
-        if random.random() < 0.3:  # 30% chance to stun
-            stun_msg = player.apply_status_effect(StatusEffectType.STUN, caster.name)
-            msg += f" {stun_msg}"
-
-        return True, msg, damage
-    return False, "", 0
-
-
-# =============================================================================
-# v5.0 New Boss Ability Handlers
-# =============================================================================
-
-def execute_ice_blast(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """AOE ice attack that damages and freezes."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        freeze_msg = player.apply_status_effect(StatusEffectType.FREEZE, caster.name)
-        return True, f"The {caster.name} unleashes an ice blast for {damage} damage! {freeze_msg}", damage
-    return False, "", 0
-
-
-def execute_freeze_ground(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Creates frozen ground that slows the player."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range + 1:
-        slow_msg = player.apply_status_effect(StatusEffectType.SLOW, caster.name)
-        return True, f"The {caster.name} freezes the ground! {slow_msg}", 0
-    return False, "", 0
-
-
-def execute_web_trap(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Immobilizes the player with sticky webs."""
-    from ..core.constants import StatusEffectType
-
-    distance = abs(player.x - caster.x) + abs(player.y - caster.y)
-    if distance <= ability.range:
-        stun_msg = player.apply_status_effect(StatusEffectType.STUN, caster.name)
-        return True, f"The {caster.name} shoots sticky webbing! {stun_msg}", 0
-    return False, "", 0
-
-
-def execute_poison_bite(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Melee attack that poisons the target."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        poison_msg = player.apply_status_effect(StatusEffectType.POISON, caster.name)
-        return True, f"The {caster.name} bites with venomous fangs for {damage} damage! {poison_msg}", damage
-    return False, "", 0
-
-
-def execute_summon_spiders(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summons spider minions."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    num_spiders = random.randint(2, 3)
-    spawned = 0
-
-    for _ in range(num_spiders):
-        for dx in range(-2, 3):
-            for dy in range(-2, 3):
-                if dx == 0 and dy == 0:
-                    continue
-                nx, ny = caster.x + dx, caster.y + dy
-                if (dungeon.is_walkable(nx, ny) and
-                    not entity_manager.get_enemy_at(nx, ny) and
-                    (nx != player.x or ny != player.y)):
-                    # Use goblin as stand-in for spider minion
-                    spider = Enemy(nx, ny, enemy_type=EnemyType.GOBLIN, is_elite=False)
-                    spider.name = "Spider"
-                    spider.symbol = 's'
-                    entity_manager.enemies.append(spider)
-                    spawned += 1
-                    break
-            if spawned >= num_spiders:
-                break
-
-    if spawned > 0:
-        return True, f"The {caster.name} summons {spawned} spiders!", 0
-    return False, "", 0
-
-
-def execute_lava_pool(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Creates a damaging lava pool (deals damage if player is close)."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        burn_msg = player.apply_status_effect(StatusEffectType.BURN, caster.name)
-        return True, f"The {caster.name} creates a pool of molten lava! {damage} damage! {burn_msg}", damage
-    return True, f"The {caster.name} creates a pool of molten lava!", 0
-
-
-def execute_inferno(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Massive AOE fire attack."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        burn_msg = player.apply_status_effect(StatusEffectType.BURN, caster.name)
-        return True, f"The {caster.name} erupts in an inferno for {damage} damage! {burn_msg}", damage
-    return False, "", 0
-
-
-def execute_summon_swarm(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summons a swarm of rat minions."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    num_rats = random.randint(3, 5)
-    spawned = 0
-
-    for _ in range(num_rats):
-        for dx in range(-2, 3):
-            for dy in range(-2, 3):
-                if dx == 0 and dy == 0:
-                    continue
-                nx, ny = caster.x + dx, caster.y + dy
-                if (dungeon.is_walkable(nx, ny) and
-                    not entity_manager.get_enemy_at(nx, ny) and
-                    (nx != player.x or ny != player.y)):
-                    # Use goblin as stand-in for rat minion
-                    rat = Enemy(nx, ny, enemy_type=EnemyType.GOBLIN, is_elite=False)
-                    rat.name = "Rat"
-                    rat.symbol = 'r'
-                    rat.health = rat.health // 2  # Weaker than goblins
-                    rat.max_health = rat.health
-                    entity_manager.enemies.append(rat)
-                    spawned += 1
-                    break
-            if spawned >= num_rats:
-                break
-
-    if spawned > 0:
-        return True, f"The {caster.name} summons a swarm of {spawned} rats!", 0
-    return False, "", 0
-
-
-def execute_plague_bite(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Diseased bite that poisons and weakens."""
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - caster.x), abs(player.y - caster.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        poison_msg = player.apply_status_effect(StatusEffectType.POISON, caster.name)
-        weak_msg = player.apply_status_effect(StatusEffectType.WEAK, caster.name)
-        return True, f"The {caster.name} delivers a plague-ridden bite for {damage} damage! {poison_msg} {weak_msg}", damage
-    return False, "", 0
-
-
-def execute_burrow(
-    ability: BossAbility,
-    caster: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Burrows underground and repositions."""
-    # Find a new position near the player
-    for _ in range(10):
-        dx = random.randint(-3, 3)
-        dy = random.randint(-3, 3)
-        nx, ny = player.x + dx, player.y + dy
-        if (dungeon.is_walkable(nx, ny) and
-            not entity_manager.get_enemy_at(nx, ny) and
-            (nx != player.x or ny != player.y)):
-            caster.x = nx
-            caster.y = ny
-            return True, f"The {caster.name} burrows underground and emerges nearby!", 0
-
-    return False, "", 0
-
-
-# =============================================================================
-# The Regent abilities (Floor 4 - Mirror Valdris)
-# =============================================================================
-
-def execute_royal_decree(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Issues a decree that summons oath-bound guards from the corners.
-
-    The Regent's signature ability - "Legitimacy by repetition."
-    Each decree summons 1-2 skeleton guards from room corners,
-    as if they were always meant to be there.
-    """
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    # Find the boss room's corners
-    current_room = dungeon.get_room_at(boss.x, boss.y)
-    if not current_room:
-        return False, "", 0
-
-    # Define corner positions within the room
-    corners = [
-        (current_room.x + 1, current_room.y + 1),  # Top-left
-        (current_room.x + current_room.width - 2, current_room.y + 1),  # Top-right
-        (current_room.x + 1, current_room.y + current_room.height - 2),  # Bottom-left
-        (current_room.x + current_room.width - 2, current_room.y + current_room.height - 2),  # Bottom-right
-    ]
-
-    num_guards = random.randint(1, 2)
-    spawned = 0
-    random.shuffle(corners)
-
-    for cx, cy in corners:
-        if spawned >= num_guards:
-            break
-        if (dungeon.is_walkable(cx, cy) and
-            not entity_manager.get_enemy_at(cx, cy) and
-            (cx != player.x or cy != player.y)):
-            # Summon oath-bound skeleton guard
-            guard = Enemy(cx, cy, enemy_type=EnemyType.SKELETON, is_elite=False)
-            guard.name = "Oath-Bound Guard"
-            guard.symbol = 'G'
-            entity_manager.enemies.append(guard)
-            spawned += 1
-
-    if spawned > 0:
-        decree_messages = [
-            "By royal decree, guards have always stood here!",
-            "The Regent rewrites reality - guards materialize!",
-            "It was always thus - guards appear from the corners!",
-        ]
-        return True, f"The Regent issues a decree: '{random.choice(decree_messages)}' ({spawned} guards summoned)", 0
-
-    return False, "", 0
-
-
-def execute_summon_guard(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """Summons a single oath-bound skeleton guard near the Regent."""
-    from .entities import Enemy
-    from ..core.constants import EnemyType
-
-    # Find spawn position near the boss
-    for dx in range(-2, 3):
-        for dy in range(-2, 3):
-            if dx == 0 and dy == 0:
-                continue
-            nx, ny = boss.x + dx, boss.y + dy
-            if (dungeon.is_walkable(nx, ny) and
-                not entity_manager.get_enemy_at(nx, ny) and
-                (nx != player.x or ny != player.y)):
-                guard = Enemy(nx, ny, enemy_type=EnemyType.SKELETON, is_elite=False)
-                guard.name = "Oath-Bound Guard"
-                guard.symbol = 'G'
-                entity_manager.enemies.append(guard)
-                return True, f"The {boss.name} summons an oath-bound guard!", 0
-
-    return False, "", 0
-
-
-def execute_counterfeit_crown(
-    ability: BossAbility,
-    boss: 'Enemy',
-    player: 'Player',
-    dungeon: 'Dungeon',
-    entity_manager: 'EntityManager',
-) -> Tuple[bool, str, int]:
-    """The Regent's crown flashes with false authority, stunning and damaging nearby.
-
-    This represents the "counterfeit reign" - the crown's power is borrowed/stolen,
-    but it still has real effects.
-    """
-    from ..core.constants import StatusEffectType
-
-    distance = max(abs(player.x - boss.x), abs(player.y - boss.y))
-    if distance <= ability.range:
-        damage = player.take_damage(ability.damage)
-        stun_msg = player.apply_status_effect(StatusEffectType.STUN, boss.name)
-        return True, f"The Regent's counterfeit crown blazes with stolen authority! {damage} damage! {stun_msg}", damage
-
-    return True, f"The Regent's crown flashes, but you're too far to be affected.", 0
+# Re-export elemental damage calculator for external use
+from .ability_handlers_elemental import calculate_elemental_damage
+
+# Import summon abilities
+from .ability_handlers_summon import (
+    execute_summon_goblins,
+    execute_raise_dead,
+    execute_raise_skeleton,
+    execute_summon_spiders,
+    execute_summon_swarm,
+    execute_royal_decree,
+    execute_summon_guard,
+)
+
+# Import elemental/magic abilities
+from .ability_handlers_elemental import (
+    execute_arcane_bolt,
+    execute_fire_breath,
+    execute_dark_bolt,
+    execute_fire_bolt,
+    execute_ice_shard,
+    execute_chain_lightning,
+    execute_ice_blast,
+    execute_freeze_ground,
+    execute_lava_pool,
+    execute_inferno,
+    execute_counterfeit_crown,
+)
+
+# Import physical/melee abilities
+from .ability_handlers_physical import (
+    execute_war_cry,
+    execute_ground_slam,
+    execute_regenerate,
+    execute_life_drain,
+    execute_teleport,
+    execute_tail_sweep,
+    execute_fire_strike,
+    execute_backstab,
+    execute_vanish,
+    execute_web_trap,
+    execute_poison_bite,
+    execute_plague_bite,
+    execute_burrow,
+)
 
 
 # Handler dispatch table mapping ability names to functions
 ABILITY_HANDLERS = {
+    # Original boss abilities (v1.0-v3.0)
     'summon_goblins': execute_summon_goblins,
     'war_cry': execute_war_cry,
     'ground_slam': execute_ground_slam,
